refactor : MovieResult 인터페이스 응집도 높이기 - MovieApi로 이동
