+++ src/apis/MovieApi.ts
@@ -1,5 +1,3 @@
-import { MovieResult } from "../../types/movieApiType";
-
 const TMDB_BASE_URL = "https://api.themoviedb.org/3";
 
 export interface MoviesResponse {
@@ -7,6 +5,23 @@ export interface MoviesResponse {
   results: MovieResult[];
   total_pages: number;
   total_results: number;
+}
+
+export interface MovieResult {
+  adult: boolean;
+  backdrop_path?: string;
+  genre_ids: number[];
+  id: number;
+  original_language: string;
+  original_title: string;
+  overview: string;
+  popularity: number;
+  poster_path: string;
+  release_date: string;
+  title: string;
+  video: boolean;
+  vote_average: number;
+  vote_count: number;
 }
 
 async function fetchWithErrorHandling(url: string): Promise<MoviesResponse> {
